Bug 2020608 - Part 15: Require unsigned integral types for CeilingLog2 and FindMostSignificantBit. r=sergesanspaille,nbp
Instead of implicitly casting the input to unsigned, require this happens explicitly
in the caller. This aligns with `<bit>` std-functions and `Rotate{Left,Right}` which
also require unsigned inputs.

Differential Revision: https://phabricator.services.mozilla.com/D286324

UltraBlame original commit: a37de510d67d8f710fd01ead6179762eacc39577

+++ js/src/jit/arm/CodeGenerator-arm.cpp
@@ -294,7 +294,7 @@ void CodeGenerator::visitMulI(LMulI* ins) {
           if (!mul->canOverflow()) {
             
             Register src = ToRegister(lhs);
-            uint32_t shift = FloorLog2(constant);
+            uint32_t shift = FloorLog2(uint32_t(constant));
             uint32_t rest = constant - (1 << shift);
             
             
@@ -320,7 +320,7 @@ void CodeGenerator::visitMulI(LMulI* ins) {
             
             
 
-            uint32_t shift = FloorLog2(constant);
+            uint32_t shift = FloorLog2(uint32_t(constant));
             if ((1 << shift) == constant) {
               
               masm.ma_lsl(Imm32(shift), ToRegister(lhs), ToRegister(dest));
@@ -406,7 +406,7 @@ void CodeGenerator::visitMulIntPtr(LMulIntPtr* ins) {
 
     
     if (constant > 0 && std::has_single_bit(uintptr_t(constant))) {
-      uint32_t shift = mozilla::FloorLog2(constant);
+      uint32_t shift = mozilla::FloorLog2(uintptr_t(constant));
       masm.ma_lsl(Imm32(shift), ToRegister(lhs), ToRegister(dest));
       return;
     }
@@ -443,7 +443,7 @@ void CodeGenerator::visitMulI64(LMulI64* lir) {
       default:
         if (constant > 0) {
           
-          int32_t shift = mozilla::FloorLog2(constant);
+          int32_t shift = mozilla::FloorLog2(uint64_t(constant));
           if (int64_t(1) << shift == constant) {
             masm.lshift64(Imm32(shift), ToRegister64(lhs));
             return;
