+++ js/src/jit/arm64/CodeGenerator-arm64.cpp
@@ -265,7 +265,7 @@ void CodeGenerator::visitMulI(LMulI* ins) {
       default:
         
         if (!mul->canOverflow() && constant > 0) {
-          int32_t shift = FloorLog2(constant);
+          int32_t shift = FloorLog2(uint32_t(constant));
           if ((1 << shift) == constant) {
             masm.Lsl(destreg32, lhsreg32, shift);
             return;
@@ -2299,7 +2299,7 @@ void CodeGenerator::visitMulI64(LMulI64* lir) {
       default:
         
         if (constant > 0) {
-          int32_t shift = mozilla::FloorLog2(constant);
+          int32_t shift = mozilla::FloorLog2(uint64_t(constant));
           if (int64_t(1) << shift == constant) {
             masm.Lsl(ARMRegister(output.reg, 64),
                      ARMRegister(ToRegister64(lhs).reg, 64), shift);
@@ -2427,7 +2427,7 @@ void CodeGenerator::visitMulIntPtr(LMulIntPtr* ins) {
 
     
     if (constant > 0 && std::has_single_bit(uintptr_t(constant))) {
-      uint32_t shift = mozilla::FloorLog2(constant);
+      uint32_t shift = mozilla::FloorLog2(uintptr_t(constant));
       masm.Lsl(dest, lhs, shift);
       return;
     }
